Formalise API to write to HKStore
It has been observed that the Health app on the iPhone is not an accurate representation of data stored.
The test that led to this conclusion was the watchOS app writing to the hk store and making an attempt to read it on the iPhone’s Health app.

To accurate dtermine whether records were written, the new `readToothBrushEvents(::)` method must be used.